+++ api_lbc.py
@@ -66,7 +66,7 @@ class LeBonCoin(object):
     ##   AUTHENTICATION   ##
     ########################
 
-    def authentication(self, force=False):
+    def authentication(self, username=None, force=False):
         """ Authentication """
         if path.isfile(self.cookie_jar_path) and not force:
             with open(self.cookie_jar_path) as cookie_jar_file:
@@ -80,7 +80,10 @@ class LeBonCoin(object):
             # Thank you, python2-3 team, for making such a fantastic mess with
             # input/raw_input :-)
             real_raw_input = vars(__builtins__).get('raw_input', input)
-            self.profile['username'] = real_raw_input('Username: ')
+            if username is None:
+                self.profile['username'] = real_raw_input('Username: ')
+            else:
+                self.profile['username'] = username
             self.profile['password'] = getpass()
             self.cookie_gen()
 
@@ -320,6 +323,8 @@ if __name__ == '__main__':
     DASHBOARD_PARSER = SUBPARSERS.add_parser('dashboard', help='List dashboard informations')
     DASHBOARD_PARSER.add_argument('--force-authentication', '-f', action='store_true',
                                   default=False, help='To force the authentication.')
+    DASHBOARD_PARSER.add_argument('--username', '-u', action='store',
+                                  default=None, help='Username to log in.')
     DASHBOARD_PARSER.add_argument('--uncolor', default=False, action='store_true',
                                   help='Disable coloration')
 
@@ -350,7 +355,7 @@ if __name__ == '__main__':
     LBC = LeBonCoin(CSV_ROOT_PATH, uncolor=ARGS.uncolor)
 
     if argv[1] == 'dashboard':
-        LBC.authentication(force=ARGS.force_authentication)
+        LBC.authentication(username=ARGS.username, force=ARGS.force_authentication)
         LBC.display_dashboard()
     elif argv[1] == 'ad':
         LBC.display_ad(ARGS.key)